Main Menu Scene, Scene Controller and Music Controller Added
The Scene controller has been coded, just have to correctly label the scene names to the ones in the script.  The music controller wont be deleted when changign scenes so any music playing will continue throughout each scene, just have to found a song and put it into the audio source attached to the music controller which is found in the main menu scene.